Use bigint for ids so the values match java's long type

+++ src/main/resources/schema.sql
@@ -1,6 +1,6 @@
 create table users
 (
-    id               int         not null primary key auto_increment,
+    id               bigint         not null primary key auto_increment,
     username         varchar(20) not null unique,
     password         varchar(64) not null,
     created_at       datetime    not null default current_timestamp,
